requirements.txt
fastapi[all]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pymongo
python-dotenv